+++ mrb_doc/core/gestures.rb
@@ -1,7 +1,0 @@
-# Returns the detected gestures based on their constant.
-# @return [Integer]
-def get_gesture_detected
-  # mrb_get_gesture_detected
-  # src/mruby_integration/core/gestures.cpp
-  GESTURE_TAP
-end