+++ src/mruby_integration/core.cpp
@@ -2,7 +2,6 @@
 
 #include "mruby_integration/core/drawing.hpp"
 #include "mruby_integration/core/files.hpp"
-#include "mruby_integration/core/gestures.hpp"
 #include "mruby_integration/core/misc.hpp"
 #include "mruby_integration/core/timing.hpp"
 
@@ -13,7 +12,6 @@ append_core(mrb_state* mrb)
 {
   append_core_drawing(mrb);
   append_core_files(mrb);
-  append_core_gestures(mrb);
   append_core_misc(mrb);
   append_core_timing(mrb);